feat: service 파일 구조 변경 - #82
