+++ src/main/java/org/mjulikelion/engnews/service/ai/AnalyzeService.java
@@ -1,8 +1,9 @@
-package org.mjulikelion.engnews.service;
+package org.mjulikelion.engnews.service.ai;
+
 
 import lombok.RequiredArgsConstructor;
-import org.mjulikelion.engnews.dto.ai.FeedbackDto;
-import org.mjulikelion.engnews.dto.ai.TrySummarizeDto;
+import org.mjulikelion.engnews.dto.ai.request.AnalyzeSentenceDto;
+import org.mjulikelion.engnews.dto.ai.response.FeedbackDto;
 import org.mjulikelion.engnews.entity.User;
 import org.springframework.http.*;
 import org.springframework.stereotype.Service;
@@ -10,11 +11,11 @@ import org.springframework.web.client.RestTemplate;
 
 @Service
 @RequiredArgsConstructor
-public class SummarizeService {
+public class AnalyzeService {
     private final RestTemplate restTemplate = new RestTemplate();
-    private final String trySummarizeURL = "http://43.203.141.103:8000/try-summarize";
+    private final String analyze_Sentence = "http://43.203.141.103:8000/analyze-sentence";
 
-    public FeedbackDto trySummarize(User user, TrySummarizeDto trySummarizeDto){
+    public FeedbackDto analyzeSentence(User user, AnalyzeSentenceDto analyzeSentenceDto){
 
         // HTTP 헤더 설정
         HttpHeaders headers = new HttpHeaders();
@@ -22,10 +23,10 @@ public class SummarizeService {
         headers.add("user", String.valueOf(user.getId()));
 
         // HTTP 요청 엔터티 생성 (헤더 + 바디)
-        HttpEntity<TrySummarizeDto> entity = new HttpEntity<>(trySummarizeDto, headers);
+        HttpEntity<AnalyzeSentenceDto> entity = new HttpEntity<>(analyzeSentenceDto, headers);
 
         // POST 요청 전송
-        ResponseEntity<String> response = restTemplate.exchange(trySummarizeURL, HttpMethod.POST, entity, String.class);
+        ResponseEntity<String> response = restTemplate.exchange(analyze_Sentence, HttpMethod.POST, entity, String.class);
 
 
         FeedbackDto feedback = FeedbackDto.builder()
